+++ docs/releases.md
@@ -5,6 +5,7 @@
 ### New Features
 
 * Nesting capability ([#204](https://github.com/CWorthy-ocean/roms-tools/pull/204))
+* Option to bypass validation ([#206](https://github.com/CWorthy-ocean/roms-tools/pull/206))
 
 ### Breaking Changes